requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.35
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
